Phase 3 Work
* Change button function

Changed button function in logincomponent so that all the data passing from the web server is in the homecomponent.

* Post and subscribe edit page

Added post and subscribe calls to the edit page buttons.

* Adding post and subscribe to settings save

Adding web server integration with get, post and subscribe calls on initialization of the page and button clicks.

* Edited Save button

Changed Save button styling to be more visible on different web-app page sizes.

* Change Settings page button

Changed styling on the settings page Save button to better fit on different web-app UI sizes.

* Logout button

Linked Logout button to the Sign-In page.

* Added input fields for settings

Added input fields to take in notification phone number and email from user or populate on load.

* Changed notification list

Changed notification location list to a multiSelect list.

* WIP

Working on sending and feeding values using the post and fetch commands.

* Fix errors

* Added to settings page load

Added to change selected elements in the dropdown menus in the settings page - on init.

* Get and post

Finished the get command and post to the web-server on the save of settings.

* Add medicine option

Added a UI element to input an new medicine name.

* Settings code clean up

Cleaned up unnecessary comments and added comments for functionalities.

* Added success label

Added success label to settings page to indicate a successful save of information.

* WIP Edit page

WIP on the edit page - adding dynamics to Save buttons

* Changed settings checkboxes

Changed the settings page multiselect to a list of checkboxes.

* Notification location checkboxes

Added notification location checkboxes to take in selected information from the user.

* Settings page label

Added label to the settings page if user enables notifications, but does not select a notification location.

* WIP

WIP - loading elements from the web server for each respective medication.

* Added pills input

Added input for the number of pills when the user is added a medication.

* Take in medication info

Transferred the data from the web-server to the web-app through the post and subscribe.

* WIP - edit and add page

WIP: loading elements into fields on the edit/add medication page

* WIP

WIP, connecting the web-server data to the overview page.

* WIP

WIP on the initialization of the add/edit page.

* Edit page OnSelect() function

Defined onSelect() function for the edit page of the application. Triggered when a medication is selected from the list in the UI.

* Add medication

Added function definition and database calls to add a medication.

* Add or Edit medication

Added definitions for methods to update/save medications.

* Edit selected medication

Changed the conditionals when checking the selected medication on the edit page.

* Overview page warnings

Added html elements and triggers to update with medication warnings on initialization of the home page.

* Dismiss notification and WIP calendar

Added function for dismiss notification and WIP on calendar functionality.

* Calendar info

Added functions and html elements to display data from return of calendar selection.

* Changed http post on calendar

Changed post command to the correct http address for the calendar command

* Formatting fixes

* Remove cache

* Add in service for repeated calls

* Code cleanups

* More cleanup

* More refactors

* Update settings.component.ts

Specified the log message from the settings page.

* Update edit.component.ts

Changed references to take array out of http received data to reflect implementation of the service to provide data.

* Code clean-up in initialization of edit component

* Added medication interface definition

Added the medication interface definition so the onSelect() and calendar select method operates correctly.

* Update calendar method

Updated calendar method definition in overview.component.ts

* Changes to edit.component.ts

Formatting fixes with indentation.

* Update edit.component.ts

Added calls to change/add data in the web-server.

* WIP

Making fixes and adding meds to the database.

* Overview component update

Added validation to ensure user selected medication before a calendar date.

* Settings component

Added label to notify user if form info has not been entered.

* Edit additions

Added labels to notify user on successful add, edit or remove of a medication. Added logic to place new data in a local Medicationx_info array.

* Update settings.component.ts

Updated data for timeForNotif to match data from the web-server.

* Fix variable typo

* Edit component additions

Edit component changes to integrate with web-server.

Co-authored-by: Carlo Villarruz <gvillarruz@ryerson.ca>